Builtin symbols : Add 't', 'f' and 'exit' as builtins
These are special because there are symbols which map directly to an
evaluable LVAL instead of mapping to a function like the others. This
means we can use these directly in functions/calls.

For example, now you just have to type 'exit' to exit the prompt,
instead of having to type 'exit ()' to force an evaluation.